Bug fix
When comparing numbers, changed `invtval` to `floatval` to properly
evaluate numbers with decimals.

+++ php/index.php
@@ -23,9 +23,9 @@ if (isset($_POST['submit'])) {
 	$body = $_POST['body'];
 	
 	// users attempt
-	$number = intval($_POST['number']);
+	$number = floatval($_POST['number']);
 	// get correct number from hidden field
-	$correctNumber = intval($_POST['correctNumber']);
+	$correctNumber = floatval($_POST['correctNumber']);
 	
 	$errors = array();
